add stefan's patch

+++ airbench/utils.py
@@ -108,7 +108,7 @@ class CifarLoader:
         self.epoch = 0
         self.images, self.labels, self.classes = data['images'], data['labels'], data['classes']
         # It's faster to load+process uint8 data than to load preprocessed fp16 data
-        self.images = (self.images.half() / 255).permute(0, 3, 1, 2).to(memory_format=torch.channels_last)
+        self.images = (self.images.float() / 255).permute(0, 3, 1, 2).to(memory_format=torch.channels_last)
 
         self.normalize = T.Normalize(CIFAR_MEAN, CIFAR_STD)
         self.proc_images = {} # Saved results of image processing to be done on the first epoch
